[f] 149 - Fetch content from supabase
