+++ src/utils/language.ts
@@ -1,0 +1,9 @@
+function getUserLanguage(): string {
+  if (navigator.language && navigator.language.startsWith('es')) {
+    return 'es'
+  }
+
+  return 'en'
+}
+
+export { getUserLanguage }
